emergency rev1d draft
stupid polygon rank screwup.  I _know_ I handled this, but somewhere it
got lost and reset, and the final gerbers had all the polys joined again
:(

seeing as we're redoing, put notches for the wings, allows both flush
with wire and through hole mounts.

Also, move the esp8266 module to the board edge to give the antenna a
better shot.