mqtt: fix will
Because async mqtt is _fulllllly_ async, it doesn't actually copy/store
the will topic and use it when you call connect, it jsut fires off the
tcp connect and tries to use the will when that actually finishes.
However, at that time, the dynamic will String had gone out of scope,
and so was considered invalid => no will.

Use a global for now.  Wills now work at least.